requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
google-genai>=1.0.0
python-dotenv==1.0.1
